requirements.txt
RPi.GPIO==0.7.1
Adafruit-WS2801==1.0.1
